rust tracelogging_dynamic - make register non-mut (#54)
* rust tracelogging_dynamic - make register non-mut

`tracelogging_dynamic::Provider` is annoying to use in multithreaded
scenarios. The pain point is that the register call takes self by `&mut`.

This is because register is the point at which the metadata is
configured. However, we don't really need to configure metadata at
registration, and in fact it probably makes more sense to configure
metadata at construction and make the object completely immutable. That
makes multithreaded use cases much nicer.

This is a breaking change, but probably worth doing. I don't think there
are a lot of users of TraceLoggingDynamic.

tracelogging\native.rs:
- I don't think there is any value in panic for overlapping calls to
  unregister, so just allow those. (We do still want to catch overlapping
  register/register or register/unregister, but we can allow overlapping
  unregister/unregister.)

tracelogging\provider.rs:
- Update doc comments with the relaxed prerequisites for unregister.

tracelogging_dynamic\provider.rs:
- Move configuration from the register method to the constructor.
- Update doc comments with the relaxed prerequisites for unregister.
- `register` now takes `self: Pin<&Self>`.
- `pin!` is now stabilized.

* Update tracelogging_dynamic version to 1.1.0

* Don't put .vscode files into git

* Comments, docs, changelog

* Use v1.1.0

---------

Co-authored-by: Doug Cook (WINDOWS) <dcook@windows.microsoft.com>