Python - use inline SHA1 implementation
TraceLoggingDynamic uses SHA1 hashing for non-security purposes.

To avoid potential problems on systems where SHA1 libraries are disabled
for security reasons, use a a local implementation of SHA1. This is
probably much slower than the library version, but it's not used in any
performance-critical code paths.

Also update some type annotations since I was getting warnings that the
`*` syntax is not supported. Seems like `Array` is as specific as you
can get with the annotations.